move module registration into PerformanceFeature class and remove bootstrap
